Fix concurrency bugs, eliminate data races, and harden preference clamping
- Make disconnect/disconnectAll async to await database cleanup instead of
  fire-and-forget Task (CalendarService, AppState)
- Remove nonisolated(unsafe) from EventScheduler by dropping the deinit that
  caused data races; use weak self in monitoring task instead
- Replace didSet re-entrant clamping in PreferencesManager with explicit
  setter methods and SwiftUI Binding accessors to eliminate infinite-loop risk
- Clamp values during loadPreferences() so corrupted UserDefaults are sanitized
- Convert FocusModeManager.readLegacyDNDPrefs to async with proper
  continuation-based process waiting instead of blocking waitUntilExit
- Fix MeetingDetailsPopupManager to not dismiss popup when a child menu is open
- Filter non-mailto EKParticipant URLs in Apple Calendar attendee conversion
- Replace JSONSerialization with Codable for Google userinfo parsing
- Use Duration-based Task.sleep throughout (SyncManager, OverlayContentView)
- Update all tests and E2E tests for new setter-based preference API

+++ Sources/Unmissable/Core/SyncManager.swift
@@ -116,7 +116,7 @@ final class SyncManager: ObservableObject {
         pendingNetworkUpdate = Task { @MainActor [weak self] in
             guard let self else { return }
             do {
-                try await Task.sleep(for: .milliseconds(Int(networkDebounceDelay * 1000)))
+                try await Task.sleep(for: .milliseconds(networkDebounceDelay * 1000))
             } catch is CancellationError {
                 return // Debounced - a newer update superseded this one
             } catch {
@@ -155,7 +155,7 @@ final class SyncManager: ObservableObject {
             // Then repeat every interval
             while !Task.isCancelled {
                 do {
-                    try await Task.sleep(for: .seconds(Int(intervalSeconds)))
+                    try await Task.sleep(for: .seconds(intervalSeconds))
                     if !Task.isCancelled {
                         await performSync()
                     }
@@ -302,7 +302,7 @@ final class SyncManager: ObservableObject {
         retryTask?.cancel()
         retryTask = Task { @MainActor in
             do {
-                try await Task.sleep(for: .seconds(Int(retryDelay)))
+                try await Task.sleep(for: .seconds(retryDelay))
                 if !Task.isCancelled {
                     await performSync()
                 }
